+++ espr/src/codegen/rust/mod.rs
@@ -1,132 +1,16 @@
 //! Generate Rust code using proc-macro utility crates
 
 mod entity;
+mod enumeration;
+mod format;
 mod rename;
 mod schema;
 mod select;
+mod simple;
+mod simple_type;
+mod type_decl;
+mod type_ref;
 
 pub use entity::*;
+pub use format::rustfmt;
 pub use schema::*;
-
-use inflector::Inflector;
-use proc_macro2::TokenStream;
-use quote::*;
-use std::{
-    io::Write,
-    process::{Command, Stdio},
-};
-
-use crate::ir::*;
-
-/// Format generated Rust code using `rustfmt` run as external process.
-///
-/// Panics
-/// -------
-/// - `rustfmt` is not found in PATH
-/// - `rustfmt` returns error
-///   - The input of rustfmt is completely generated by this crate,
-///     and thus it must be a bug, i.e. not a runtime error.
-///
-pub fn rustfmt(tt: String) -> String {
-    let mut child = Command::new("rustfmt")
-        .stdin(Stdio::piped())
-        .stdout(Stdio::piped())
-        .spawn()
-        .expect("Failed to spawn rustfmt process");
-
-    // Write input from another thread for avoiding deadlock.
-    // See https://doc.rust-lang.org/std/process/index.html#handling-io
-    let mut stdin = child.stdin.take().expect("Failed to open stdin");
-    std::thread::spawn(move || {
-        stdin
-            .write_all(tt.as_bytes())
-            .expect("Failed to write to stdin");
-    });
-    let output = child
-        .wait_with_output()
-        .expect("Failed to wait output of rustfmt process");
-
-    // non-UTF8 comment should be handled in the tokenize phase,
-    // and not be included in IR.
-    String::from_utf8(output.stdout).expect("rustfmt output contains non-UTF8 input")
-}
-
-impl ToTokens for Simple {
-    fn to_tokens(&self, tokens: &mut TokenStream) {
-        let id = format_ident!("{}", &self.id.to_pascal_case());
-        let ty = &self.ty;
-        tokens.append_all(quote! {
-            #[derive(Clone, Debug, PartialEq, AsRef, Deref, DerefMut, ::serde::Serialize, ::serde::Deserialize)]
-            pub struct #id(pub #ty);
-        });
-    }
-}
-
-impl ToTokens for Enumeration {
-    fn to_tokens(&self, tokens: &mut TokenStream) {
-        let id = format_ident!("{}", &self.id.to_pascal_case());
-        let items: Vec<_> = self
-            .items
-            .iter()
-            .map(|i| format_ident!("{}", i.to_pascal_case()))
-            .collect();
-        tokens.append_all(quote! {
-            #[derive(Debug, Clone, PartialEq, ::serde::Deserialize)]
-            pub enum #id {
-                #( #items ),*
-            }
-        });
-    }
-}
-
-impl ToTokens for TypeDecl {
-    fn to_tokens(&self, tokens: &mut TokenStream) {
-        match self {
-            TypeDecl::Simple(simple) => simple.to_tokens(tokens),
-            TypeDecl::Rename(rename) => rename.to_tokens(tokens),
-            TypeDecl::Enumeration(e) => e.to_tokens(tokens),
-            TypeDecl::Select(select) => select.to_tokens(tokens),
-        }
-    }
-}
-
-impl ToTokens for SimpleType {
-    fn to_tokens(&self, tokens: &mut TokenStream) {
-        use crate::ast::SimpleType::*;
-        match self.0 {
-            Number => tokens.append(format_ident!("f64")),
-            Real => tokens.append(format_ident!("f64")),
-            Integer => tokens.append(format_ident!("i64")),
-            Logical => tokens.append_all(quote! { Logical }),
-            Boolen => tokens.append(format_ident!("bool")),
-            String_ { .. } => tokens.append(format_ident!("String")),
-            Binary { .. } => unimplemented!("Binary type is not supported yet"),
-        }
-    }
-}
-
-impl ToTokens for TypeRef {
-    fn to_tokens(&self, tokens: &mut TokenStream) {
-        use TypeRef::*;
-        match self {
-            SimpleType(ty) => ty.to_tokens(tokens),
-            Named { name, .. } => {
-                let name = format_ident!("{}", name.to_pascal_case());
-                tokens.append_all(quote! { #name });
-            }
-            Entity {
-                name, is_supertype, ..
-            } => {
-                let name = if *is_supertype {
-                    format_ident!("{}Any", name.to_pascal_case())
-                } else {
-                    format_ident!("{}", name.to_pascal_case())
-                };
-                tokens.append_all(quote! { #name });
-            }
-            Set { base, .. } | List { base, .. } => {
-                tokens.append_all(quote! { Vec<#base> });
-            }
-        }
-    }
-}
