+++ espr/src/codegen/rust/type_decl.rs
@@ -1,0 +1,15 @@
+use proc_macro2::TokenStream;
+use quote::*;
+
+use crate::ir::*;
+
+impl ToTokens for TypeDecl {
+    fn to_tokens(&self, tokens: &mut TokenStream) {
+        match self {
+            TypeDecl::Simple(simple) => simple.to_tokens(tokens),
+            TypeDecl::Rename(rename) => rename.to_tokens(tokens),
+            TypeDecl::Enumeration(e) => e.to_tokens(tokens),
+            TypeDecl::Select(select) => select.to_tokens(tokens),
+        }
+    }
+}
